### 2025-10-20 6e3150a
chore: prepare for PyOpenSci submission
- Fix version consistency (1.0.0) across pyproject.toml and CITATION.cff
- Update contact email from example.com to ang@hezhiang.com
- Clean up temporary files and build artifacts
- Remove duplicate python_wrapper directory
- Improve .gitignore with comprehensive exclusions
- Fix test cases referencing deleted directories
- Rename submission_patterns to manuscript and ignore it
- Remove PRIVACY.md and Chinese CHANGELOG.md
- Ensure all core code and documentation use English
- All 44 tests passing successfully

This commit ensures the project meets PyOpen
[…]
