+++ crates/langlint_translators/src/google.rs
@@ -1,0 +1,401 @@
+//! Google Translate translator using the free API
+
+use crate::{Translator, TranslationError, TranslationResult};
+use async_trait::async_trait;
+use rand::Rng;
+use serde::Deserialize;
+use std::collections::HashMap;
+use std::time::Duration;
+use tokio::time::sleep;
+
+/// Configuration for Google translator
+#[derive(Debug, Clone)]
+pub struct GoogleConfig {
+    /// Timeout for requests in seconds
+    pub timeout: u64,
+    /// Retry count for failed requests
+    pub retry_count: u32,
+    /// Random delay range in milliseconds (to avoid rate limiting)
+    pub delay_range: (u64, u64),
+    /// Custom service URLs (if any)
+    pub service_urls: Option<Vec<String>>,
+}
+
+impl Default for GoogleConfig {
+    fn default() -> Self {
+        Self {
+            timeout: 30,
+            retry_count: 3,
+            delay_range: (300, 600), // 300-600ms to respect rate limits
+            service_urls: None,
+        }
+    }
+}
+
+/// Google Translate API response
+#[derive(Debug, Deserialize)]
+#[allow(dead_code)]
+struct GoogleTranslateResponse {
+    #[serde(rename = "sentences")]
+    sentences: Option<Vec<Sentence>>,
+    #[serde(rename = "src")]
+    detected_language: Option<String>,
+}
+
+#[derive(Debug, Deserialize)]
+#[allow(dead_code)]
+struct Sentence {
+    trans: Option<String>,
+    orig: Option<String>,
+}
+
+/// Google Translator implementation
+pub struct GoogleTranslator {
+    config: GoogleConfig,
+    client: reqwest::Client,
+    language_mapping: HashMap<String, String>,
+}
+
+impl GoogleTranslator {
+    /// Create a new Google translator with default config
+    pub fn new() -> Result<Self, TranslationError> {
+        Self::with_config(GoogleConfig::default())
+    }
+
+    /// Create a new Google translator with custom config
+    pub fn with_config(config: GoogleConfig) -> Result<Self, TranslationError> {
+        let client = reqwest::Client::builder()
+            .timeout(Duration::from_secs(config.timeout))
+            .build()
+            .map_err(|e| TranslationError::NetworkError(e))?;
+
+        let mut language_mapping = HashMap::new();
+        
+        // Add supported languages (subset for now, can be expanded)
+        let languages = [
+            ("en", "english"),
+            ("zh", "chinese (simplified)"),
+            ("zh-cn", "chinese (simplified)"),
+            ("zh-tw", "chinese (traditional)"),
+            ("ja", "japanese"),
+            ("ko", "korean"),
+            ("fr", "french"),
+            ("de", "german"),
+            ("es", "spanish"),
+            ("it", "italian"),
+            ("pt", "portuguese"),
+            ("ru", "russian"),
+            ("ar", "arabic"),
+            ("hi", "hindi"),
+            ("th", "thai"),
+            ("vi", "vietnamese"),
+            ("id", "indonesian"),
+            ("nl", "dutch"),
+            ("sv", "swedish"),
+            ("da", "danish"),
+            ("no", "norwegian"),
+            ("fi", "finnish"),
+            ("pl", "polish"),
+            ("tr", "turkish"),
+            ("cs", "czech"),
+            ("hu", "hungarian"),
+            ("ro", "romanian"),
+            ("bg", "bulgarian"),
+            ("el", "greek"),
+            ("he", "hebrew"),
+            ("uk", "ukrainian"),
+        ];
+
+        for (code, name) in &languages {
+            language_mapping.insert(code.to_string(), name.to_string());
+        }
+
+        Ok(Self {
+            config,
+            client,
[…]
