+++ langlint/parsers/generic_code_parser.py
@@ -1,0 +1,576 @@
+"""
+Generic code parser for extracting comments from various programming languages.
+
+This parser uses pre-compiled, high-performance regular expressions to identify
+single-line and block comments in various programming languages.
+"""
+
+import re
+from typing import List, Optional, Dict, Any, Tuple
+from dataclasses import dataclass
+
+from .base import Parser, TranslatableUnit, ParseResult, UnitType, ParseError
+
+
+@dataclass
+class CommentPattern:
+    """Represents a comment pattern for a programming language."""
+    single_line: List[str]
+    block_start: Optional[str] = None
+    block_end: Optional[str] = None
+    block_nested: bool = False
+
+
+class GenericCodeParser(Parser):
+    """
+    Generic parser for various programming languages that extracts comments.
+
+    This parser uses pre-compiled regular expressions to identify and extract
+    single-line and block comments from various programming languages.
+    """
+
+    # Language-specific comment patterns
+    COMMENT_PATTERNS = {
+        '.js': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.ts': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.go': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.rs': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.java': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.cpp': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.c': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.h': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.cs': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.php': CommentPattern(
+            single_line=['//', '#'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.rb': CommentPattern(
+            single_line=['#'],
+            block_start='=begin',
+            block_end='=end'
+        ),
+        '.sh': CommentPattern(
+            single_line=['#']
+        ),
+        '.bash': CommentPattern(
+            single_line=['#']
+        ),
+        '.zsh': CommentPattern(
+            single_line=['#']
+        ),
+        '.fish': CommentPattern(
+            single_line=['#']
+        ),
+        '.sql': CommentPattern(
+            single_line=['--'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.r': CommentPattern(
+            single_line=['#']
+        ),
+        '.m': CommentPattern(
+            single_line=['%'],
+            block_start='%{',
+            block_end='%}'
+        ),
+        '.scala': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.kt': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.swift': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.dart': CommentPattern(
+            single_line=['//'],
+            block_start='/*',
+            block_end='*/'
+        ),
+        '.lua': CommentPattern(
+            single_line=['--'],
+            block_start='--[[',
[…]
