+++ crates/langlint_core/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod types;
+pub mod config;
+pub mod cache;
+
+// Re-export commonly used types
+pub use types::{ParseResult, Position, Priority, TranslatableUnit, UnitType};
+pub use config::Config;
+pub use cache::Cache;
