+++ CITATION.cff
@@ -1,9 +1,9 @@
 cff-version: 1.2.0
 message: "If you use this software, please cite it as below."
 authors:
-  - family-names: "LangLint Team"
-    given-names: ""
-    orcid: "https://orcid.org/0000-0000-0000-0000"
+  - family-names: "He"
+    given-names: "Zhiang"
+    email: "ang@hezhiang.com"
 title: "LangLint: A scalable, domain-agnostic platform for automated translation and standardization of structured text in scientific collaboration"
 version: 1.0.0
 date-released: 2024-01-01
@@ -24,8 +24,9 @@ preferred-citation:
   type: software
   title: "LangLint: A scalable, domain-agnostic platform for automated translation and standardization of structured text in scientific collaboration"
   authors:
-    - family-names: "LangLint Team"
-      given-names: ""
+    - family-names: "He"
+      given-names: "Zhiang"
+      email: "ang@hezhiang.com"
   year: 2024
   url: "https://github.com/HzaCode/Langlint"
   license: MIT
