+++ pyproject.toml
@@ -6,8 +6,8 @@ build-backend = "setuptools.build_meta"
 name = "langlint"
 version = "1.0.0"
 description = "A scalable, domain-agnostic platform for automated translation and standardization of structured text in scientific collaboration"
-authors = [{name = "LangLint Team", email = "ang@hezhiang.com"}]
-maintainers = [{name = "LangLint Team", email = "ang@hezhiang.com"}]
+authors = [{name = "Zhiang He", email = "ang@hezhiang.com"}]
+maintainers = [{name = "Zhiang He", email = "ang@hezhiang.com"}]
 license = {text = "MIT"}
 readme = "README.md"
 requires-python = ">=3.8"
