+++ Cargo.toml
@@ -10,7 +10,7 @@ resolver = "2"
 
 [workspace.package]
 version = "1.0.0"
-authors = ["Langlint Team"]
+authors = ["Zhiang He <ang@hezhiang.com>"]
 edition = "2021"
 license = "MIT"
 repository = "https://github.com/HzaCode/Langlint"
