### 2025-10-20 6e78f40
Add automatic documentation deployment workflow
