### 2025-10-20 f06147c
docs: restore comprehensive Sphinx documentation
- Restore docs/ directory with all RST files
- Includes API documentation, CLI docs, configuration guide
- Ready for Sphinx build and deployment
