+++ docs/requirements.txt
@@ -1,0 +1,5 @@
+sphinx>=7.0.0
+sphinx-rtd-theme>=2.0.0
+myst-parser>=2.0.0
+sphinx-autodoc-typehints>=1.24.0
+
