+++ docs/usage.rst
@@ -1,0 +1,293 @@
+Usage Guide
+===========
+
+This guide covers the main features and usage patterns of LangLint.
+
+Supported Languages
+-------------------
+
+LangLint supports 100+ language pairs, including:
+
+European Languages
+~~~~~~~~~~~~~~~~~~
+
+* English (en)
+* French (fr)
+* German (de)
+* Spanish (es)
+* Italian (it)
+* Portuguese (pt)
+* Russian (ru)
+* Dutch (nl)
+* Polish (pl)
+* Swedish (sv)
+
+Asian Languages
+~~~~~~~~~~~~~~~
+
+* Simplified Chinese (zh-CN)
+* Traditional Chinese (zh-TW)
+* Japanese (ja)
+* Korean (ko)
+* Thai (th)
+* Vietnamese (vi)
+* Hindi (hi)
+* Indonesian (id)
+
+Other Languages
+~~~~~~~~~~~~~~~
+
+* Arabic (ar)
+* Hebrew (he)
+* Turkish (tr)
+* Greek (el)
+* Persian (fa)
+
+Supported File Types
+--------------------
+
+LangLint supports 28+ file types:
+
+Programming Languages
+~~~~~~~~~~~~~~~~~~~~~
+
+* **Python**: ``.py``
+* **JavaScript/TypeScript**: ``.js``, ``.ts``, ``.jsx``, ``.tsx``
+* **Rust**: ``.rs``
+* **Go**: ``.go``
+* **C/C++**: ``.c``, ``.cpp``, ``.h``, ``.hpp``
+* **Java**: ``.java``
+* **Kotlin**: ``.kt``
+* **Scala**: ``.scala``
+* **C#**: ``.cs``
+* **PHP**: ``.php``
+* **Ruby**: ``.rb``
+* **Swift**: ``.swift``
+* **Dart**: ``.dart``
+* **Lua**: ``.lua``
+
+Documentation & Data
+~~~~~~~~~~~~~~~~~~~~
+
+* **Markdown**: ``.md``
+* **Jupyter Notebooks**: ``.ipynb``
+* **Shell Scripts**: ``.sh``, ``.bash``
+* **SQL**: ``.sql``
+* **R**: ``.r``, ``.R``
+* **MATLAB**: ``.m``
+* **Vim Script**: ``.vim``
+
+What Gets Translated
+--------------------
+
+LangLint intelligently translates:
+
+✅ **Translated**:
+  * Comments (``# inline comments``)
+  * Docstrings (``"""docstring"""``)
+  * Documentation strings
+  * Markdown text content
+
+❌ **Not Translated** (Preserved):
+  * Variable names
+  * Function names
+  * String literals
+  * Code syntax
+  * Import statements
+  * Configuration values
+
+Command Options
+---------------
+
+scan Command
+~~~~~~~~~~~~
+
+.. code-block:: bash
+
+   langlint scan [OPTIONS] PATH
+
+Options:
+
+* ``-o, --output FILE`` - Save results to file
+* ``--format FORMAT`` - Output format (json/yaml/csv)
+* ``-i, --include PATTERN`` - Include file patterns
+* ``-e, --exclude PATTERN`` - Exclude file patterns
+* ``-v, --verbose`` - Enable verbose output
+
+Examples:
+
+.. code-block:: bash
+
+   # Scan with JSON output
+   langlint scan src/ -o report.json --format json
+
+   # Scan with exclusions
+   langlint scan . -e "**/test_*" -e "**/__pycache__/"
+
+translate Command
+~~~~~~~~~~~~~~~~~
+
+.. code-block:: bash
+
+   langlint translate [OPTIONS] PATH
+
+Options:
+
+* ``-s, --source-lang LANG`` - Source language code (default: auto)
[…]
