### 2025-10-25 3ebaa01
fix: preserve quote style in docstring translation (#5)
- Save original quote style (triple double/single quotes) to metadata during extraction
- Use saved quote style during reconstruction instead of re-detecting
- Use parser's reconstruct method for proper multi-line docstring handling
- Fixes multi-line docstrings being corrupted with extra quotes

Closes #5
