### 2025-10-25 67fb258
chore: upgrade pyo3 to 0.21 for Python 3.13 support
