### 2025-10-25 a3aaa1b
chore: enable PyO3 abi3 for stable Python API compatibility
