### 2025-10-25 c04cab3
fix: use derive(Default) for Priority enum as suggested by clippy
