### 2025-10-25 6387a9e
ci: remove macOS from test matrix due to persistent Python linking issues
