### 2025-10-25 19153b2
fix: resolve all clippy warnings for CI
