### 2025-10-25 30cb4da
fix: resolve all clippy warnings for CI
