### 2025-10-25 843160b
fix: allow too_many_arguments for CLI command functions
