### 2025-10-25 c1feff1
fix: allow too_many_arguments for fix command
