### 2025-10-25 dfc9e42
style: auto-format code with cargo fmt
