+++ crates/langlint_cli/src/commands/mod.rs
@@ -1,6 +1,5 @@
 //! CLI command implementations
 
+pub mod fix;
 pub mod scan;
 pub mod translate;
-pub mod fix;
-
