+++ docs_source/api/core.rst
@@ -1,0 +1,29 @@
+Core Module
+===========
+
+The core module provides the fundamental components of LangLint.
+
+Dispatcher
+----------
+
+.. automodule:: langlint.core.dispatcher
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Configuration
+-------------
+
+.. automodule:: langlint.core.config
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Cache
+-----
+
+.. automodule:: langlint.core.cache
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
