### 2025-10-25 7304a17
style: auto-format code with cargo fmt
