+++ crates/langlint_parsers/src/python.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use langlint_core::{ParseResult, TranslatableUnit, UnitType, Priority};
+use langlint_core::{ParseResult, Priority, TranslatableUnit, UnitType};
 use regex::Regex;
 use std::sync::OnceLock;
 
@@ -16,7 +16,7 @@ impl PythonParser {
     /// Check if text should be translated
     fn is_translatable(&self, text: &str) -> bool {
         let text = text.trim();
-        
+
         // Skip empty or very short text
         if text.len() < 3 {
             return false;
@@ -30,25 +30,27 @@ impl PythonParser {
         // Skip text that's mostly code or symbols
         // Count alphabetic chars AND CJK characters (Chinese, Japanese, Korean)
         let char_count = text.chars().count();
-        let meaningful_count = text.chars().filter(|c| {
-            c.is_alphabetic() || 
+        let meaningful_count = text
+            .chars()
+            .filter(|c| {
+                c.is_alphabetic() ||
             ('\u{4E00}'..='\u{9FFF}').contains(c) ||  // CJK Unified Ideographs
             ('\u{3400}'..='\u{4DBF}').contains(c) ||  // CJK Extension A
             ('\u{3040}'..='\u{30FF}').contains(c) ||  // Hiragana + Katakana
-            ('\u{AC00}'..='\u{D7AF}').contains(c)     // Hangul
-        }).count();
-        
+            ('\u{AC00}'..='\u{D7AF}').contains(c) // Hangul
+            })
+            .count();
+
         if meaningful_count < char_count / 3 {
             return false;
         }
 
         // Skip common technical terms
         let technical_terms = [
-            "TODO", "FIXME", "NOTE", "HACK", "XXX",
-            "self", "cls", "args", "kwargs",
-            "return", "def", "class", "import",
+            "TODO", "FIXME", "NOTE", "HACK", "XXX", "self", "cls", "args", "kwargs", "return",
+            "def", "class", "import",
         ];
-        
+
         if technical_terms.contains(&text) {
             return false;
         }
@@ -74,7 +76,11 @@ impl Parser for PythonParser {
 
     fn can_parse(&self, path: &str, content: Option<&str>) -> bool {
         // Check extension
-        if self.supported_extensions().iter().any(|ext| path.ends_with(ext)) {
+        if self
+            .supported_extensions()
+            .iter()
+            .any(|ext| path.ends_with(ext))
+        {
             return true;
         }
 
@@ -97,15 +103,13 @@ impl Parser for PythonParser {
 
         // Regex for docstrings (simplified)
         static DOCSTRING_RE: OnceLock<Regex> = OnceLock::new();
-        let docstring_re = DOCSTRING_RE.get_or_init(|| {
-            Regex::new(r#"^\s*["']{3}(.+?)["']{3}"#).unwrap()
-        });
-        
+        let docstring_re =
+            DOCSTRING_RE.get_or_init(|| Regex::new(r#"^\s*["']{3}(.+?)["']{3}"#).unwrap());
+
         // Regex for multi-line docstring start
         static DOCSTRING_START_RE: OnceLock<Regex> = OnceLock::new();
-        let docstring_start_re = DOCSTRING_START_RE.get_or_init(|| {
-            Regex::new(r#"^\s*(["']{3})\s*(.*)$"#).unwrap()
-        });
+        let docstring_start_re =
+            DOCSTRING_START_RE.get_or_init(|| Regex::new(r#"^\s*(["']{3})\s*(.*)$"#).unwrap());
 
         let mut i = 0;
         while i < lines.len() {
@@ -117,14 +121,10 @@ impl Parser for PythonParser {
[…]
