+++ test_issue5_exact.py
@@ -1,0 +1,9 @@
+def func():
+    """
+    Line 1
+    Line 2
+    Line 3
+    """
+    pass
+
+
