+++ crates/langlint_core/src/lib.rs
@@ -1,8 +1,8 @@
-pub mod types;
-pub mod config;
 pub mod cache;
+pub mod config;
+pub mod types;
 
 // Re-export commonly used types
-pub use types::{ParseResult, Position, Priority, TranslatableUnit, UnitType};
-pub use config::Config;
 pub use cache::Cache;
+pub use config::Config;
+pub use types::{ParseResult, Position, Priority, TranslatableUnit, UnitType};
