+++ crates/langlint_cli/src/commands/fix.rs
@@ -24,20 +24,23 @@ pub async fn execute(
 ) -> Result<()> {
     // Load config to get backup preference (if not overridden by CLI)
     let config = Config::find_and_load().unwrap_or_default();
-    
+
     // Command line --no-backup flag overrides config file
     let should_backup = if no_backup {
-        false  // Explicit --no-backup flag
+        false // Explicit --no-backup flag
     } else {
-        config.backup  // Use config file setting (default: true)
+        config.backup // Use config file setting (default: true)
     };
-    
+
     if verbose {
         println!("{} {}", "Fixing (in-place translate):".bold().cyan(), path);
         println!("  Source language: {}", source);
         println!("  Target language: {}", target);
         println!("  Translator: {}", translator_name);
-        println!("  Backup: {}", if should_backup { "enabled" } else { "disabled" });
+        println!(
+            "  Backup: {}",
+            if should_backup { "enabled" } else { "disabled" }
+        );
     }
 
     // Create translator
@@ -52,10 +55,10 @@ pub async fn execute(
     }
 
     let path_obj = Path::new(path);
-    
+
     // Collect files to translate
     let files = collect_files(path_obj)?;
-    
+
     if files.is_empty() {
         println!("{} No translatable files found", "!".yellow());
         return Ok(());
@@ -68,11 +71,17 @@ pub async fn execute(
     // Confirm before proceeding
     if !yes && !files.is_empty() {
         if should_backup {
-            println!("\n{} About to translate {} files in-place (backups will be created)", 
-                "⚠".yellow(), files.len());
+            println!(
+                "\n{} About to translate {} files in-place (backups will be created)",
+                "⚠".yellow(),
+                files.len()
+            );
         } else {
-            println!("\n{} About to translate {} files in-place (⚠️  NO BACKUP will be created)", 
-                "⚠".yellow(), files.len());
+            println!(
+                "\n{} About to translate {} files in-place (⚠️  NO BACKUP will be created)",
+                "⚠".yellow(),
+                files.len()
+            );
         }
         println!("  Press Enter to continue, Ctrl+C to cancel...");
         let mut input = String::new();
@@ -96,7 +105,16 @@ pub async fn execute(
         let filename = file_path.file_name().unwrap().to_string_lossy();
         pb.set_message(format!("Translating {}", filename));
 
-        match translate_file(file_path, source, target, translator.as_ref(), should_backup, verbose).await {
+        match translate_file(
+            file_path,
+            source,
+            target,
+            translator.as_ref(),
+            should_backup,
+            verbose,
+        )
+        .await
+        {
             Ok(units) => {
                 if units > 0 {
                     translated_count += 1;
@@ -137,11 +155,15 @@ pub async fn execute(
     }
 
     if should_backup {
-        println!("\n{} Translation complete! Backups created with .backup extension", 
-            "✓".green().bold());
+        println!(
+            "\n{} Translation complete! Backups created with .backup extension",
+            "✓".green().bold()
+        );
     } else {
-        println!("\n{} Translation complete! (No backups created)", 
-            "✓".green().bold());
+        println!(
+            "\n{} Translation complete! (No backups created)",
+            "✓".green().bold()
+        );
     }
 
     Ok(())
@@ -178,7 +200,9 @@ async fn translate_file(
     }
 
     // Translate all units
-    let texts: Vec<String> = parse_result.units.iter()
+    let texts: Vec<String> = parse_result
+        .units
+        .iter()
         .map(|u| u.content.clone())
         .collect();
 
@@ -197,15 +221,15 @@ async fn translate_file(
[…]

[…]
