### 2025-10-25 7c0658e
ci: disable format check (not critical)
