### 2025-10-25 34d2a06
ci: update all actions to latest versions (v3->v4, v4->v5)
