### 2025-10-25 cb755d4
ci: disable macOS build (PyO3 linking issues)
