### 2025-10-25 b16cff5
feat: add non-ASCII filter to Rust parsers and comprehensive tests (Issue #4)
