+++ crates/langlint_cli/Cargo.toml
@@ -20,3 +20,6 @@ tokio = { version = "1.0", features = ["full"] }
 colored = "2.1"
 indicatif = "0.17"
 walkdir = "2.5"
+
+[dev-dependencies]
+tempfile = "3.8"
