### 2025-10-25 0df9560
Fix issue #3: Add directory translation support for Rust CLI
- Add collect_files() function to recursively scan directories
- Add should_translate() to filter supported file types
- Add translate_single_file() for individual file processing
- Support output to directory with preserved structure
- Add automatic backup creation for in-place translation
- Add progress bar for batch translation
- Add 6 unit tests for new functionality
- Add walkdir and tempfile dependencies

Fixes #3
