+++ crates/langlint_cli/src/commands/translate.rs
@@ -7,7 +7,8 @@ use langlint_core::ParseResult;
 use langlint_parsers::{GenericCodeParser, NotebookParser, Parser, PythonParser};
 use langlint_translators::{GoogleTranslator, MockTranslator, Translator};
 use std::fs;
-use std::path::Path;
+use std::path::{Path, PathBuf};
+use walkdir::WalkDir;
 
 /// Execute the translate command
 #[allow(clippy::too_many_arguments)]
@@ -42,141 +43,281 @@ pub async fn execute(
         println!("{} Translator created", "✓".green());
     }
 
-    // Read file
     let path_obj = Path::new(path);
-    let content =
-        fs::read_to_string(path_obj).with_context(|| format!("Failed to read file: {}", path))?;
 
-    if verbose {
-        println!("{} File loaded ({} bytes)", "✓".green(), content.len());
-    }
+    // Collect files to translate
+    let files = collect_files(path_obj)?;
 
-    // Parse file to extract translatable units
-    let parse_result = parse_file(path, &content)?;
-    let unit_count = parse_result.units.len();
-
-    if unit_count == 0 {
-        println!("{} No translatable units found", "!".yellow());
+    if files.is_empty() {
+        println!("{} No translatable files found", "!".yellow());
         return Ok(());
     }
 
-    println!("{} Found {} translatable units", "✓".green(), unit_count);
+    if verbose {
+        println!("{} {} files found", "Total:".bold(), files.len());
+    }
 
-    // Translate units (with parallel processing for better performance)
-    let progress = if !verbose && !dry_run {
-        let pb = ProgressBar::new(unit_count as u64);
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("{spinner:.green} [{bar:40.cyan/blue}] {pos}/{len} {msg}")
-                .unwrap()
-                .progress_chars("#>-"),
-        );
-        Some(pb)
-    } else {
-        None
-    };
+    // Setup progress bar
+    let pb = ProgressBar::new(files.len() as u64);
+    pb.set_style(
+        ProgressStyle::default_bar()
+            .template("{spinner:.green} [{bar:40.cyan/blue}] {pos}/{len} {msg}")
+            .unwrap()
+            .progress_chars("=>-"),
+    );
 
-    let mut translated_units = Vec::new();
-    for (i, unit) in parse_result.units.iter().enumerate() {
-        if let Some(ref pb) = progress {
-            pb.set_message(format!("Translating unit {}/{}", i + 1, unit_count));
-        }
+    let mut translated_count = 0;
+    let mut error_count = 0;
+    let mut total_units = 0;
 
-        if verbose {
-            println!(
-                "\n{} Translating unit {}/{}: {:?}",
-                "→".blue(),
-                i + 1,
-                unit_count,
-                unit.unit_type
-            );
-            let preview = if unit.content.len() > 60 {
-                format!("{}...", &unit.content[..60])
+    for file_path in &files {
+        let filename = file_path.file_name().unwrap().to_string_lossy();
+        pb.set_message(format!("Translating {}", filename));
+
+        // Determine output path for this file
+        let output_file_path = if let Some(output_dir) = output {
+            // Calculate relative path from input to maintain directory structure
+            let relative_path = if path_obj.is_dir() {
+                file_path.strip_prefix(path_obj).unwrap_or(file_path)
             } else {
-                unit.content.clone()
+                file_path.file_name().map(Path::new).unwrap_or(file_path)
             };
-            println!("  Original: {}", preview.dimmed());
-        }
+            PathBuf::from(output_dir).join(relative_path)
+        } else {
+            file_path.clone()
+        };
 
-        if !dry_run {
-            match translator.translate(&unit.content, source, target).await {
-                Ok(result) => {
[…]
