+++ langlint/cli.py
@@ -28,39 +28,88 @@ from .translators.mock_translator import MockTranslator, MockConfig
 
 # Global console instance
 # Use force_terminal=True and legacy_windows=False to avoid gbk encoding issues on Windows
-console = Console(force_terminal=True, legacy_windows=False)
+# Force UTF-8 output for better compatibility
+import sys
+console = Console(force_terminal=True, legacy_windows=False, file=sys.stderr)
 
 
 @click.group()
 @click.version_option(version="1.0.0", prog_name="langlint")
-@click.option('--verbose', '-v', is_flag=True, help='Enable verbose output')
 @click.option('--config', '-c', type=click.Path(exists=True), help='Path to configuration file')
 @click.pass_context
-def cli(ctx: click.Context, verbose: bool, config: Optional[str]) -> None:
+def cli(ctx: click.Context, config: Optional[str]) -> None:
     """
     LangLint: A scalable, domain-agnostic platform for automated translation
     and standardization of structured text in scientific collaboration.
     """
     ctx.ensure_object(dict)
-    ctx.obj['verbose'] = verbose
     ctx.obj['config'] = config
 
 
 @cli.command()
 @click.argument('path', type=click.Path(exists=True))
 @click.option('--output', '-o', type=click.Path(), help='Output file for results')
-@click.option('--format', 'output_format', type=click.Choice(['json', 'yaml', 'csv']), default='json', help='Output format')
+@click.option('--format', '-f', 'output_format', type=click.Choice(['json', 'yaml', 'csv', 'text']), default='json', help='Output format')
+@click.option('--verbose', '-v', is_flag=True, help='Enable verbose output')
 @click.option('--include', '-i', multiple=True, help='Include patterns')
 @click.option('--exclude', '-e', multiple=True, help='Exclude patterns')
 @click.pass_context
-def scan(ctx: click.Context, path: str, output: Optional[str], output_format: str, include: tuple, exclude: tuple) -> None:
+def scan(ctx: click.Context, path: str, output: Optional[str], output_format: str, verbose: bool, include: tuple, exclude: tuple) -> None:
     """
     Scan files for translatable text without translating.
 
     PATH: Path to file or directory to scan
     """
     try:
-        # Load configuration
+        # Try to use Rust implementation for better performance
+        try:
+            import langlint_py
+            import json
+            
+            # Prepare exclude patterns
+            exclude_list = list(exclude) if exclude else []
+            
+            # Call Rust implementation
+            result_json = langlint_py.scan(
+                path,
+                format=output_format,
+                verbose=verbose,
+                exclude=exclude_list
+            )
+            
+            # Output results
+            if output:
+                Path(output).write_text(result_json, encoding='utf-8')
+                print(f"Results written to {output}")
+            else:
+                # Output based on format
+                if output_format == 'json':
+                    print(result_json)
+                elif output_format == 'text':
+                    # For text format, check if result is JSON or plain text
+                    try:
+                        result = json.loads(result_json)
+                        # Display nicely for text format
+                        print(f"\nFiles scanned: {result['files_scanned']}")
+                        print(f"Total units: {result['total_units']}")
+                        
+                        if verbose:
+                            for file_result in result.get('results', []):
+                                print(f"  {file_result['file']}: {file_result['units']} units")
+                    except json.JSONDecodeError:
+                        # Already in text format from Rust
+                        print(result_json)
+                else:
+                    # For yaml, csv etc., output JSON (can be converted later)
+                    print(result_json)
+            
+            return
+            
+        except ImportError:
+            # Fall back to Python implementation
[…]
