+++ test_dir/file1.py
@@ -1,6 +1,0 @@
-# 文件1
-def hello():
-    """问候函数"""
-    print("你好")
-
-
