+++ test_final_comprehensive.py
@@ -1,28 +1,0 @@
-# -*- coding: utf-8 -*-
-"""模块文档
-多行描述
-"""
-
-class Calculator:
-    """计算器类
-    提供基本数学运算
-    """
-    
-    def add(self, a, b):
-        """加法函数
-        返回两数之和
-        """
-        # 执行加法
-        return a + b
-    
-    def subtract(self, a, b):
-        '''减法函数
-        返回两数之差
-        '''
-        return a - b
-    
-    def multiply(self, a, b):
-        """乘法"""
-        return a * b
-
-
