+++ check_consistency.py
@@ -1,201 +1,0 @@
-#!/usr/bin/env python3
-"""
-LangLint 一致性检查脚本
-检查文档、代码和README之间的一致性
-"""
-
-import os
-import re
-import sys
-from pathlib import Path
-from typing import Dict, List, Tuple, Any
-import json
-
-class ConsistencyChecker:
-    def __init__(self, project_root: str = "."):
-        self.project_root = Path(project_root)
-        self.issues = []
-        
-    def check_file_exists(self, file_path: str) -> bool:
-        """检查文件是否存在"""
-        return (self.project_root / file_path).exists()
-    
-    def read_file(self, file_path: str) -> str:
-        """读取文件内容"""
-        try:
-            with open(self.project_root / file_path, 'r', encoding='utf-8') as f:
-                return f.read()
-        except Exception as e:
-            self.issues.append(f"无法读取文件 {file_path}: {e}")
-            return ""
-    
-    def check_backup_extension_consistency(self) -> List[str]:
-        """检查备份文件扩展名一致性"""
-        issues = []
-        
-        # 检查README.md
-        readme_content = self.read_file("README.md")
-        if ".bak" in readme_content:
-            issues.append("README.md 中使用了 .bak 扩展名，应该使用 .backup")
-        
-        # 检查docs目录中的RST文件
-        docs_files = [
-            "docs/cli.rst",
-            "docs/quickstart.rst", 
-            "docs/configuration.rst"
-        ]
-        
-        for doc_file in docs_files:
-            if self.check_file_exists(doc_file):
-                content = self.read_file(doc_file)
-                if ".bak" in content:
-                    issues.append(f"{doc_file} 中使用了 .bak 扩展名，应该使用 .backup")
-        
-        return issues
-    
-    def check_config_file_consistency(self) -> List[str]:
-        """检查配置文件一致性"""
-        issues = []
-        
-        # 检查README中提到的配置文件
-        readme_content = self.read_file("README.md")
-        
-        # 检查是否提到了正确的配置文件
-        if ".langlint.yml" not in readme_content:
-            issues.append("README.md 中缺少 .langlint.yml 配置文件说明")
-        
-        if "langlint.toml" not in readme_content:
-            issues.append("README.md 中缺少 langlint.toml 配置文件说明")
-        
-        return issues
-    
-    def check_cli_options_consistency(self) -> List[str]:
-        """检查CLI选项一致性"""
-        issues = []
-        
-        # 检查Python CLI
-        cli_py_content = self.read_file("langlint/cli.py")
-        if "--no-backup" not in cli_py_content:
-            issues.append("Python CLI (langlint/cli.py) 缺少 --no-backup 选项")
-        
-        # 检查Rust CLI
-        rust_cli_content = self.read_file("crates/langlint_cli/src/main.rs")
-        if "--no-backup" not in rust_cli_content:
-            issues.append("Rust CLI 缺少 --no-backup 选项")
-        
-        return issues
-    
-    def check_test_coverage_consistency(self) -> List[str]:
-        """检查测试覆盖率一致性"""
-        issues = []
-        
-        readme_content = self.read_file("README.md")
-        
-        # 检查README中的测试数量
-        test_count_match = re.search(r'(\d+)\+?\s*tests?', readme_content)
-        if test_count_match:
-            reported_count = int(test_count_match.group(1))
-            # 这里可以添加实际测试数量的检查
-            pass
-        
-        # 检查覆盖率百分比
-        coverage_match = re.search(r'(\d+)%\s*coverage', readme_content)
-        if coverage_match:
-            reported_coverage = int(coverage_match.group(1))
-            # 这里可以添加实际覆盖率的检查
-            pass
-        
[…]
