+++ test_dir/file2.py
@@ -1,6 +1,0 @@
-# 文件2
-def goodbye():
-    """告别函数"""
-    print("再见")
-
-
