### 2025-10-25 851a5e8
refactor: remove unimplemented OpenAI and DeepL translator support
