+++ langlint/translators/openai_translator.py
@@ -1,551 +1,0 @@
-"""
-OpenAI translator for using GPT models for translation.
-
-This translator uses OpenAI's GPT models to provide high-quality translations
-with context awareness and natural language understanding.
-"""
-
-import asyncio
-import json
-from typing import List, Dict, Any, Optional
-from dataclasses import dataclass
-
-try:
-    import openai
-    from openai import AsyncOpenAI
-except ImportError:
-    openai = None
-    AsyncOpenAI = None
-
-from .base import Translator, TranslationResult, TranslationStatus, TranslationError
-
-
-@dataclass
-class OpenAIConfig:
-    """Configuration for OpenAI translator."""
-    api_key: str
-    model: str = "gpt-3.5-turbo"
-    max_tokens: int = 4000
-    temperature: float = 0.3
-    max_retries: int = 3
-    timeout: int = 30
-    base_url: Optional[str] = None
-
-
-class OpenAITranslator(Translator):
-    """
-    Translator using OpenAI's GPT models for translation.
-
-    This translator provides high-quality translations with context awareness
-    and natural language understanding using OpenAI's GPT models.
-    """
-
-    # Language code mapping for OpenAI
-    LANGUAGE_MAPPING = {
-        'en': 'English',
-        'zh': 'Chinese',
-        'ja': 'Japanese',
-        'ko': 'Korean',
-        'fr': 'French',
-        'de': 'German',
-        'es': 'Spanish',
-        'it': 'Italian',
-        'pt': 'Portuguese',
-        'ru': 'Russian',
-        'ar': 'Arabic',
-        'hi': 'Hindi',
-        'th': 'Thai',
-        'vi': 'Vietnamese',
-        'id': 'Indonesian',
-        'ms': 'Malay',
-        'tl': 'Filipino',
-        'tr': 'Turkish',
-        'pl': 'Polish',
-        'nl': 'Dutch',
-        'sv': 'Swedish',
-        'da': 'Danish',
-        'no': 'Norwegian',
-        'fi': 'Finnish',
-        'cs': 'Czech',
-        'hu': 'Hungarian',
-        'ro': 'Romanian',
-        'bg': 'Bulgarian',
-        'hr': 'Croatian',
-        'sk': 'Slovak',
-        'sl': 'Slovenian',
-        'et': 'Estonian',
-        'lv': 'Latvian',
-        'lt': 'Lithuanian',
-        'uk': 'Ukrainian',
-        'be': 'Belarusian',
-        'mk': 'Macedonian',
-        'sq': 'Albanian',
-        'sr': 'Serbian',
-        'bs': 'Bosnian',
-        'me': 'Montenegrin',
-        'is': 'Icelandic',
-        'ga': 'Irish',
-        'cy': 'Welsh',
-        'mt': 'Maltese',
-        'eu': 'Basque',
-        'ca': 'Catalan',
-        'gl': 'Galician',
-        'af': 'Afrikaans',
-        'sw': 'Swahili',
-        'am': 'Amharic',
-        'az': 'Azerbaijani',
-        'bn': 'Bengali',
-        'gu': 'Gujarati',
-        'he': 'Hebrew',
-        'ka': 'Georgian',
-        'kk': 'Kazakh',
-        'ky': 'Kyrgyz',
-        'lo': 'Lao',
-        'mn': 'Mongolian',
-        'my': 'Burmese',
-        'ne': 'Nepali',
-        'si': 'Sinhala',
-        'ta': 'Tamil',
-        'te': 'Telugu',
-        'ur': 'Urdu',
-        'uz': 'Uzbek',
-    }
-
-    def __init__(self, config: OpenAIConfig):
-        """
-        Initialize the OpenAI translator.
-
-        Args:
-            config: OpenAI configuration
-        """
-        if openai is None:
[…]
