+++ langlint/translators/deepl_translator.py
@@ -1,395 +1,0 @@
-"""
-DeepL translator for using DeepL API for translation.
-
-This translator uses DeepL's translation API to provide high-quality translations
-with support for many languages and specialized terminology.
-"""
-
-import asyncio
-from typing import List, Dict, Any, Optional
-from dataclasses import dataclass
-
-try:
-    import deepl
-    from deepl import Translator as DeepLTranslator
-except ImportError:
-    deepl = None
-    DeepLTranslator = None
-
-from .base import Translator, TranslationResult, TranslationStatus, TranslationError
-
-
-@dataclass
-class DeepLConfig:
-    """Configuration for DeepL translator."""
-    api_key: str
-    base_url: Optional[str] = None
-    max_retries: int = 3
-    timeout: int = 30
-    formality: str = "default"  # default, more, less, prefer_more, prefer_less
-
-
-class DeepLTranslator(Translator):
-    """
-    Translator using DeepL's translation API.
-
-    This translator provides high-quality translations with support for many
-    languages and specialized terminology using DeepL's translation service.
-    """
-
-    # Language code mapping for DeepL
-    LANGUAGE_MAPPING = {
-        'en': 'EN',
-        'de': 'DE',
-        'fr': 'FR',
-        'it': 'IT',
-        'ja': 'JA',
-        'es': 'ES',
-        'nl': 'NL',
-        'pl': 'PL',
-        'pt': 'PT',
-        'ru': 'RU',
-        'zh': 'ZH',
-        'ko': 'KO',
-        'ar': 'AR',
-        'bg': 'BG',
-        'cs': 'CS',
-        'da': 'DA',
-        'el': 'EL',
-        'et': 'ET',
-        'fi': 'FI',
-        'hu': 'HU',
-        'id': 'ID',
-        'lt': 'LT',
-        'lv': 'LV',
-        'nb': 'NB',
-        'ro': 'RO',
-        'sk': 'SK',
-        'sl': 'SL',
-        'sv': 'SV',
-        'tr': 'TR',
-        'uk': 'UK',
-    }
-
-    def __init__(self, config: DeepLConfig):
-        """
-        Initialize the DeepL translator.
-
-        Args:
-            config: DeepL configuration
-        """
-        if deepl is None:
-            raise ImportError("deepl is required for DeepL translation. Install with: pip install deepl")
-
-        self.config = config
-        self.client = DeepLTranslator(
-            auth_key=config.api_key,
-            server_url=config.base_url
-        )
-
-    @property
-    def name(self) -> str:
-        """Return the name of this translator."""
-        return "DeepL"
-
-    @property
-    def supported_languages(self) -> List[str]:
-        """Return a list of supported language codes."""
-        return list(self.LANGUAGE_MAPPING.keys())
-
-    async def translate(
-        self,
-        text: str,
-        source_language: str,
-        target_language: str,
-        **kwargs: Any
-    ) -> TranslationResult:
-        """
-        Translate text from source language to target language.
-
-        Args:
-            text: Text to translate
-            source_language: Source language code
-            target_language: Target language code
-            **kwargs: Additional parameters
-
-        Returns:
-            TranslationResult containing the translated text and metadata
-
-        Raises:
-            TranslationError: If translation fails
-        """
-        # Validate languages
-        self.validate_languages(source_language, target_language)
-
-        # Normalize language codes
-        source_lang = self.normalize_language_code(source_language)
-        target_lang = self.normalize_language_code(target_language)
-
-        # Get DeepL language codes
-        source_deepl = self.LANGUAGE_MAPPING.get(source_lang)
-        target_deepl = self.LANGUAGE_MAPPING.get(target_lang)
-
[…]
