### 2025-10-25 1801226
docs: add PRIVACY.md with comprehensive privacy policy
