### 2025-10-25 be528f4
docs: add privacy section to README with link to PRIVACY.md
