### 2025-10-25 4f47ee2
docs: add comprehensive examples for Python translation and Jupyter notebooks
