### 2025-10-25 339c7e5
fix: remove OpenAI and DeepL imports from translators module
