+++ langlint/translators/__init__.py
@@ -1,13 +1,10 @@
 """
 Translation modules for different translation services.
 
-This package contains translators for various translation APIs including
-OpenAI, DeepL, Google Translate, and Azure Translator.
+This package contains translators for Google Translate and mock testing.
 """
 
 from .base import Translator, TranslationResult, TranslationError
-from .openai_translator import OpenAITranslator
-from .deepl_translator import DeepLTranslator
 from .google_translator import GoogleTranslator
 from .mock_translator import MockTranslator
 
@@ -15,8 +12,6 @@ __all__ = [
     "Translator",
     "TranslationResult",
     "TranslationError",
-    "OpenAITranslator",
-    "DeepLTranslator",
     "GoogleTranslator",
     "MockTranslator",
 ]
