### 2025-10-25 d2b51f1
fix: remove OpenAI and DeepL imports from CLI module
