+++ _sources/api/parsers.rst.txt
@@ -1,0 +1,53 @@
+Parsers Module
+==============
+
+The parsers module provides file type-specific parsers for extracting translatable content.
+
+Base Parser
+-----------
+
+.. automodule:: langlint.parsers.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Python Parser
+-------------
+
+.. automodule:: langlint.parsers.python_parser
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Markdown Parser
+---------------
+
+.. automodule:: langlint.parsers.markdown_parser
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Notebook Parser
+---------------
+
+.. automodule:: langlint.parsers.notebook_parser
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Generic Code Parser
+-------------------
+
+.. automodule:: langlint.parsers.generic_code_parser
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Config Parser
+-------------
+
+.. automodule:: langlint.parsers.config_parser
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
