+++ _sources/api/translators.rst.txt
@@ -1,0 +1,36 @@
+Translators
+===========
+
+This module contains translator implementations for different translation services.
+
+Base Translator
+---------------
+
+.. automodule:: langlint.translators.base
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Google Translator
+-----------------
+
+.. automodule:: langlint.translators.google_translator
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Mock Translator
+---------------
+
+.. automodule:: langlint.translators.mock_translator
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Future Translators
+------------------
+
+The following translators are planned for future releases:
+
+* OpenAI GPT (requires Rust implementation)
+* DeepL API (requires Rust implementation)

[…]
