### 2025-10-25 10ede40
refactor: simplify Python package to thin CLI wrapper calling Rust binary
