+++ langlint/parsers/python_parser.py
@@ -1,700 +1,0 @@
-"""
-Python parser for extracting translatable text from Python files.
-
-This parser uses Python's built-in tokenize module to precisely extract
-comments and docstrings while preserving the exact structure of the code.
-"""
-
-import tokenize
-import io
-from typing import List, Optional, Dict, Any
-from dataclasses import dataclass
-
-from .base import Parser, TranslatableUnit, ParseResult, UnitType, ParseError
-
-
-@dataclass
-class PythonToken:
-    """Represents a Python token with position information."""
-    type: int
-    string: str
-    start: tuple[int, int]  # (line, column)
-    end: tuple[int, int]
-    line: str
-
-
-class PythonParser(Parser):
-    """
-    Parser for Python files that extracts comments and docstrings.
-
-    This parser uses Python's tokenize module to precisely identify
-    and extract translatable text while maintaining exact positioning
-    information for reconstruction.
-    """
-
-    @property
-    def supported_extensions(self) -> List[str]:
-        """Return supported file extensions."""
-        return ['.py', '.pyi', '.pyw']
-
-    @property
-    def supported_mime_types(self) -> List[str]:
-        """Return supported MIME types."""
-        return [
-            'text/x-python',
-            'application/x-python-code',
-            'text/x-python-script'
-        ]
-
-    def can_parse(self, file_path: str, content: Optional[str] = None) -> bool:
-        """
-        Check if this parser can handle the given file.
-
-        Args:
-            file_path: Path to the file
-            content: Optional file content
-
-        Returns:
-            True if this parser can handle the file
-        """
-        if content is None:
-            return file_path.endswith(tuple(self.supported_extensions))
-
-        # Simple heuristic: check for Python-specific patterns
-        sample = content[:500].strip()
-
-        # Must have Python-specific patterns
-        python_patterns = [
-            'def ', 'class ', 'import ', 'from ', 'if ', 'for ', 'while ',
-            'return ', 'yield ', 'async ', 'await ', 'try:', 'except:',
-            'finally:', 'with ', 'lambda ', 'and ', 'or ', 'not ',
-            'in ', 'is ', '==', '!=', '<=', '>=', '=', '+=', '-=',
-            '**', '//', '%', '&', '|', '^', '~', '<<', '>>'
-        ]
-
-        has_python_patterns = any(pattern in sample for pattern in python_patterns)
-
-        # Additional check: if it's just plain text, reject
-        if not has_python_patterns:
-            # Check if it looks like plain text (no Python syntax)
-            lines = [line.strip() for line in sample.split('\n') if line.strip()]
-            if len(lines) > 1:
-                # If multiple lines but no Python syntax, likely not Python
-                return False
-
-        return has_python_patterns
-
-    def extract_translatable_units(self, content: str, file_path: str) -> ParseResult:
-        """
-        Extract translatable units from Python file content.
-
-        Args:
-            content: The file content to parse
-            file_path: Path to the file
-
-        Returns:
-            ParseResult containing all translatable units found
-
-        Raises:
-            ParseError: If the file cannot be parsed
-        """
-        try:
-            # Tokenize the content
-            tokens = self._tokenize_content(content)
-
-            # Extract translatable units
-            units = []
-            line_count = content.count('\n') + 1
-
-            for token in tokens:
-                unit = self._create_translatable_unit(token, content)
-                if unit:
-                    units.append(unit)
-
-            return ParseResult(
-                units=units,
-                file_type="python",
-                encoding="utf-8",
-                line_count=line_count,
-                metadata={
-                    "parser": "PythonParser",
-                    "version": "1.0.0"
-                }
-            )
-
-        except Exception as e:
[…]
