### 2025-10-25 e890f16
docs: remove Jupyter notebook support and update language examples
