### 2025-10-25 98c22e0
test: remove legacy Python implementation tests, keep only Rust binding tests
