+++ examples/python_translation/example_japanese.py
@@ -1,31 +1,20 @@
-"""
-データ処理ユーティリティ
-"""
+"""データ处理ユーティリティ"""
 
 
 def process_data(data):
-    """
-    データを処理する
-    
-    引数:
-        data: 処理するデータ
-    
-    戻り値:
-        処理されたデータ
-    """
-    # データを検証
+    """データを处理する 引数: 数据: 处理するデータ 戻り値: 处理されたデータ"""
+    # データを検证
     if not data:
         return None
     
-    # データを変換
+    # データを変换
     result = [x * 2 for x in data]
     return result
 
 
 def validate_input(value):
-    """入力値を検証する"""
+    """入力値を検证する"""
     # 値が正の数かチェック
     if value > 0:
         return True
     return False
-

[…]
