### 2025-10-25 7ee346f
test: add comprehensive tests achieving 97% coverage
