### 2025-10-25 cbd75ef
ci: enable CI workflow with Python and Rust tests, update README badges
