### 2025-10-25 b52d489
docs: simplify release description
