### 2025-10-25 ed7b1f1
chore: remove Jupyter notebook examples
