### 2025-10-25 74713a5
docs: remove version release statement
