### 2025-10-25 187ee6e
fix: resolve clippy warning in translate command
