### 2025-10-25 ad64754
docs: add documentation passing badge
