### 2025-10-25 3c90ae4
docs: sync documentation with README and standardize to zh-CN
- Add documentation badges (ReadTheDocs link)
- Update all examples to use zh-CN instead of zh for consistency
- Sync basic usage examples between README and docs
- Update docs/index.rst with full badge list (Rust, CI, codecov, Downloads)
- Standardize command examples across all documentation files
