### 2025-10-25 38003c6
docs: fix documentation link to GitHub Pages
