### 2025-10-25 064157c
test: add timeout decorators to prevent CI hangs
- Add @pytest.mark.timeout decorators to long-running tests
- Fix test_scan_command_execution to use existing example file
- Add pytest-timeout>=2.1.0 to dev dependencies
- Set 10s timeout for scan tests, 15s for translate/fix tests
- Prevents exit code 143 (SIGTERM) in CI environment
- Maintains full code coverage (no mocking)
