### 2025-10-25 f0d0050
test: register timeout marker in pytest.ini
- Add timeout marker to pytest configuration
- Fixes 'timeout' not found in markers configuration error
- Required for pytest-timeout plugin to work with --strict-markers
