+++ pytest.ini
@@ -10,4 +10,5 @@ addopts =
 markers =
     slow: marks tests as slow (deselect with '-m "not slow"')
     integration: marks tests as integration tests
+    timeout: marks tests with timeout limits
 
