### 2025-10-25 9e2bb57
style: fix all ruff linting errors and add Ruff badge
- Fix E731: Replace lambda with def function in __init__.py
- Fix F401: Remove all unused imports in tests
- Fix E722: Replace bare except with except Exception
- Fix E712: Use pythonic True/False checks (not X, X instead of == False/True)
- Add Ruff badge to README
- All 21 tests passing with 97% coverage
- ruff check passes with no errors
