### 2025-10-25 2bb6041
fix: prevent OOM in CI tests and bump version to 1.0.1
- Skip resource-intensive tests in CI environment to avoid OOM killer (exit code 137)
- Use --help commands in CI for test_scan_command_execution
- Skip test_scan_with_real_file, test_translate_with_mock, test_fix_with_mock in CI
- Bump version from 1.0.0 to 1.0.1 across all config files
- Updated pyproject.toml, Cargo.toml, and __init__.py
