+++ crates/langlint_py/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "maturin"
 
 [project]
 name = "langlint"
-version = "1.0.0"
+version = "1.0.1"
 description = "High-performance automated translation platform powered by Rust"
 authors = [{name = "Zhiang He", email = "ang@hezhiang.com"}]
 license = {text = "MIT"}
