### 2025-10-25 fd82a5d
test: update version tests to 1.0.1 and skip Rust tests in CI
- Update expected version from 1.0.0 to 1.0.1 in all tests
- Skip test_has_rust_true when Rust module not available
- Fixes CI test failures related to missing Rust extension
