+++ tests/test_import_fallback.py
@@ -25,7 +25,7 @@ def test_import_error_fallback():
             assert not langlint.HAS_RUST
             assert langlint.scan is None
             assert langlint.translate is None
-            assert langlint.version() == "1.0.0"
+            assert langlint.version() == "1.0.1"
                 
     finally:
         # Restore original modules
@@ -36,5 +36,8 @@ def test_import_error_fallback():
 def test_has_rust_true():
     """Test that HAS_RUST is True when module is available"""
     import langlint
+    # Skip if Rust module is not available (e.g., in CI without Rust build)
+    if not langlint.HAS_RUST:
+        pytest.skip("Rust module not available")
     assert langlint.HAS_RUST
 
