### 2025-10-25 fcf1eee
ci: use maturin to build and install Rust extension for full test coverage
- Replace 'cargo build' with 'maturin develop' to properly install Rust extension
- Add pytest-timeout to CI dependencies
- This fixes coverage dropping from 97% to 73% in CI
- Now HAS_RUST will be True in CI environment
- Enables full test execution with Rust bindings
