### 2025-10-25 0c4a38e
ci: fix maturin build - use 'build' instead of 'develop' for CI
- Replace 'maturin develop' with 'maturin build' (doesn't require virtualenv)
- Add step to install built wheel with pip
- Fixes 'Couldn't find a virtualenv' error in CI
