### 2025-10-25 0b037f7
ci: add debug output and fix wheel installation path
