### 2025-10-25 6955cb8
ci: add force-reinstall and import verification debug
