### 2025-10-25 47da86e
fix: correct maturin module-name to langlint_py (top-level)
- Change module-name from 'langlint.langlint_py' to 'langlint_py'
- Add wheel format to include paths for Python files
- This should fix 'ModuleNotFoundError: No module named langlint_py' in CI
