### 2025-10-25 108f99f
ci: force bash shell for cross-platform compatibility
- Add 'shell: bash' to steps using bash commands (ls, find, etc)
- Fixes Windows CI failure with PowerShell incompatibility
