### 2025-10-25 2e366d1
feat: add keywords to pyproject.toml for better package discoverability
