### 2025-10-26 d9ce014
Revise README for clarity and emphasis on accessibility
Updated the project description to emphasize accessibility and international collaboration in scientific software. Enhanced the overview and key features sections for clarity.
