### 2025-12-09 eabf885
Add awesome-i18n badge linking to Python section
