### 2025-12-16 #268 @eliotwrobson
@HzaCode sorry for the delay here. Because of high submission volume, we have instituted a new policy around multiple submissions: https://www.pyopensci.org/software-peer-review/our-process/policies.html#submission-volume-and-maintainer-overlap

In keeping with this, could you designate one submission to be the primary one to focus on for now among your current ones? I think we should go with the one that has the most github stars so far #267, but it is up to you. Once we hear back, I'll mark all of the other submissions as on hold. Please let us know if you have any questions. Thanks!
