### 2025-12-16 #268 @HzaCode
Hi @eliotwrobson,

Thank you for the update!

I completely understand the new policy regarding submission volume. To be honest, it is precisely because of pyOpenSci's high standards and rigorous peer-review process that I was so eager to submit my work here. I see this platform as a benchmark for quality in the ecosystem, so I am happy to follow the process that ensures the best review experience for everyone.

Let’s proceed with OneCite (#267) as my single active submission. You made the right call—it is the tool I personally use most frequently for my own research, so it makes sense to start there.

Please mark this submission (and any others) as on hold for now. I look forward to focusing on the OneCite review!
