### 2025-12-16 d407ac8
docs: add Featured in Awesome Static Analysis badge
