### 2025-12-17 591ed37
chore: bump version to 1.0.3 and fix maturin packaging
