+++ crates/langlint_py/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "langlint_py"
-version = "1.0.1"
+version.workspace = true
 edition = "2021"
 
 [lib]
