+++ crates/langlint_translators/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "langlint_translators"
-version = "1.0.0"
+version.workspace = true
 edition = "2021"
 
 [dependencies]
