+++ .pre-commit-config-test.yaml
@@ -26,3 +26,6 @@ repos:
 
 
 
+
+
+

[…]
