### 2025-12-17 2205570
fix: remove readme path that breaks sdist build
