### 2026-03-02 #268 @crhea93
@HzaCode
I want  to start by thanking you for the submission of your code. After much internal deliberation, we have decided not to continue with the submission at this time and will be closing the issue presently. We have determined that this work does not sufficiently fit within our community's parameters for utility or design.
